top: mipi core: give delay to the valid as there can be fifo in core
In here, there could be the FIFO in the MIPI RX & TX Core given
from the Efinix, supposing that the data will be delayed automatically
while receiving and transmitting of a clock cycle.

Therefore, data is always coming back from TX as pixels are not
properly oriented.

To remove the problem, here, valid will be delayed for 2 clock
cycles as it is getting delayed while receiving as well as
transmitting.

Also removed, fixed comments and removed unnecessary declarations
from the design.